--- clone_tiggul_t6/src/App.js
@@ -1,8 +1,0 @@
-import React from "react";
-import Router from "./shared/Router";
-
-function App() {
-  return <Router />;
-}
-
-export default App;
