--- clone_tiggul_t6/src/components/matchState.jsx
@@ -1,1 +1,0 @@
-//최근 일주일 매칭 현황
