--- clone_tiggul_t6/src/elements/buttons.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Button = () => {
-  return (
-    <>
-      <div></div>
-    </>
-  );
-};
-
-export default Button;
