--- clone_tiggul_t6/src/index.js
@@ -1,20 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import { Provider } from "react-redux";
-import App from "./App";
-import store from "./redux/config/configStore";
-import reportWebVitals from "./reportWebVitals";
-import GlobalStyle from "./style/globalstyle";
-
-const root = ReactDOM.createRoot(document.getElementById("root"));
-root.render(
-  <Provider store={store}>
-    <GlobalStyle />
-    <App />
-  </Provider>
-);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
