--- clone_tiggul_t6/src/pages/mypage.jsx
@@ -1,22 +1,0 @@
-import React from "react";
-import { useNavigate } from "react-router-dom";
-
-const MyPage = () => {
-  const navigate = useNavigate();
-  return (
-    <>
-      <div>
-        {" "}
-        <button
-          onClick={() => {
-            navigate("/");
-          }}
-        >
-          홈으로!!!
-        </button>
-      </div>
-    </>
-  );
-};
-
-export default MyPage;
