--- clone_tiggul_t6/src/pages/partyleader/leaderfive.jsx
@@ -1,21 +1,0 @@
-import React from "react";
-import { useNavigate } from "react-router-dom";
-
-const LeaderFive = () => {
-  const navigate = useNavigate();
-  return (
-    <>
-      <div>
-        <button
-          onClick={() => {
-            navigate("/");
-          }}
-        >
-          홈으로!!!
-        </button>
-      </div>
-    </>
-  );
-};
-
-export default LeaderFive;
