+++ src/index.js
@@ -1,15 +1,17 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
+import { Provider } from "react-redux";
 import App from "./App";
+import store from "./redux/config/configStore";
 import reportWebVitals from "./reportWebVitals";
 import GlobalStyle from "./style/globalstyle";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
-  <React.StrictMode>
+  <Provider store={store}>
     <GlobalStyle />
     <App />
-  </React.StrictMode>
+  </Provider>
 );
 
 // If you want to start measuring performance in your app, pass a function
