+++ src/pages/week7-teeggul.code-workspace
@@ -1,0 +1,8 @@
+{
+  "folders": [
+	{
+		"path": "../.."
+	}
+],
+  "settings": {}
+}
